- Add new vuex store module for model. - Move model code from homecards-mixin.vue into model.js. - Modify model loading / buildPathInModel to avoid circular references in parent and modelPath - call loadSemanticModel after loadSemantics in app.vue - replace calls to loadModel() with calls to loadSemanticModel in home.vue and home-edit.vue
