+++ bundles/org.openhab.ui/web/src/components/app.vue
@@ -484,7 +484,8 @@ export default {
               ? [this.$oh.api.get('/rest/ui/components/ui:page'), this.$oh.api.get('/rest/ui/components/ui:widget')]
               : [Promise.resolve([]), Promise.resolve([])],
             dayjsLocalePromise,
-            this.$store.dispatch('loadSemantics')
+            this.$store.dispatch('loadSemantics'),
+            this.$store.dispatch('loadSemanticModel')
           ])
         }).then((data) => {
           // store the pages & widgets
